+++ mypyc/emitclass.py
@@ -5,7 +5,7 @@ from typing import Optional, List, Tuple, Dict, Callable, Mapping, Set
 from collections import OrderedDict
 
 from mypyc.common import PREFIX, NATIVE_PREFIX, REG_PREFIX
-from mypyc.emit import Emitter, HeaderDeclaration
+from mypyc.codegen.emit import Emitter, HeaderDeclaration
 from mypyc.emitfunc import native_function_header, native_getter_name, native_setter_name
 from mypyc.emitwrapper import (
     generate_dunder_wrapper, generate_hash_wrapper, generate_richcompare_wrapper,
