+++ mypyc/emitwrapper.py
@@ -5,7 +5,7 @@ from typing import List, Optional
 from mypy.nodes import ARG_POS, ARG_OPT, ARG_NAMED_OPT, ARG_NAMED, ARG_STAR, ARG_STAR2
 
 from mypyc.common import PREFIX, NATIVE_PREFIX, DUNDER_PREFIX
-from mypyc.emit import Emitter
+from mypyc.codegen.emit import Emitter
 from mypyc.ir.rtypes import (
     RType, is_object_rprimitive, is_int_rprimitive, is_bool_rprimitive, object_rprimitive
 )
