+++ mypyc/test/test_emit.py
@@ -2,7 +2,7 @@ import unittest
 
 from mypy.nodes import Var
 
-from mypyc.emit import Emitter, EmitterContext
+from mypyc.codegen.emit import Emitter, EmitterContext
 from mypyc.ir.ops import BasicBlock, Environment
 from mypyc.ir.rtypes import int_rprimitive
 from mypyc.namegen import NameGenerator
