+++ mypyc/test/test_emitwrapper.py
@@ -3,7 +3,7 @@ from typing import List
 
 from mypy.test.helpers import assert_string_arrays_equal
 
-from mypyc.emit import Emitter, EmitterContext
+from mypyc.codegen.emit import Emitter, EmitterContext
 from mypyc.emitwrapper import generate_arg_check
 from mypyc.ir.rtypes import list_rprimitive, int_rprimitive
 from mypyc.namegen import NameGenerator
